+++ Makefile.in.mk
@@ -1,55 +1,4 @@
 # -----------------------------------------------
-# independent variables
-
-DIR_VENV := $(shell pipenv --venv 2>/dev/null)
-
-
-# -----------------------------------------------
-# OS-depend variables
-
-ifeq ($(OS), Windows_NT)
-
-DIR_REPO := $(shell cd)
-
-else
-
-DIR_REPO := $(shell pwd)
-
-endif
-
-
-# -----------------------------------------------
-# Paths
-
-DIR_SCRIPTS = $(DIR_REPO)/scripts
-DIR_SRC := $(DIR_REPO)/src
-DIR_TESTS := $(DIR_REPO)/tests
-
-# -----------------------------------------------
-# Virtualenv-depend variables
-
-ifeq ($(shell python "$(DIR_SCRIPTS)/detect_venv.py"), True)
-
-IN_VENV := True
-RUN :=
-PIPENV_INSTALL := echo Cannot create venv under venv
-
-else
-
-IN_VENV := False
-RUN := pipenv run
-PIPENV_INSTALL := pipenv install
-
-endif
-
-
-# -----------------------------------------------
-# calculated variables
-
-PYTHON := $(RUN) python
-
-
-# -----------------------------------------------
 # OS-depend actions
 
 ifeq ($(OS), Windows_NT)
@@ -69,3 +18,56 @@ define log
 endef
 
 endif
+
+
+# -----------------------------------------------
+# independent variables
+
+DIR_VENV := $(shell $(PIPENV) --venv 2>/dev/null)
+DIR_REPO := $(realpath ./)
+
+# -----------------------------------------------
+# OS-depend variables
+
+ifeq ($(OS), Windows_NT)
+
+DIR_REPO := $(abspath $(shell cd))
+
+else
+
+DIR_REPO := $(abspath $(shell pwd))
+
+endif
+
+
+# -----------------------------------------------
+# Paths
+
+DIR_SRC := $(abspath $(DIR_REPO)/src)
+DIR_TESTS := $(abspath $(DIR_REPO)/tests)
+
+DIR_SCRIPTS = $(abspath $(DIR_SRC)/scripts)
+
+
+# -----------------------------------------------
+# Virtualenv-depend variables
+
+ifeq ($(shell python "$(DIR_SRC)/framework/util/venv.py"), True)
+
+IN_VENV := True
+RUN :=
+PIPENV_INSTALL := echo Cannot create venv under venv
+
+else
+
+IN_VENV := False
+RUN := $(PIPENV) run
+PIPENV_INSTALL := $(PIPENV) install
+
+endif
+
+
+# -----------------------------------------------
+# calculated variables
+
+PYTHON := $(RUN) python
