+++ Pipfile
@@ -19,4 +19,4 @@ gunicorn = "*"
 psycopg2-binary = "*"
 
 [requires]
-python_version = "3.8"
+python_version = "3.9"
