+++ config/settings.yml
@@ -1,5 +1,4 @@
 default:
-  ALLOWED_HOSTS: [ ]
   DATABASE_URL: ""
   DIRS_EXCLUDED:
     - .idea
@@ -7,24 +6,27 @@ default:
     - .tests_artifacts
   DIRS_RESOURCES:
     - config
-    - scripts
   DIRS_SOURCES:
     - src
   DIRS_TEMPLATES: [ ]
+  HOST: ""
   MODE_DEBUG: true
   MODE_PROFILING: false
   PORT: -1
   PROJECT_NAME: ""
+  SENTRY_DSN: ""
   TEMPLATE_ENGINE: "Django"
   VENV_SYNTHETIC: false
   WEB_CONCURRENCY: 0
 
 
 development:
+  HOST: "localhost"
   PORT: 8000
 
 
 ci:
+  HOST: "localhost"
   PORT: 8000
   SECRET_KEY: "1"
   WEB_CONCURRENCY: 2
