--- src/framework/consts.py
@@ -1,9 +1,0 @@
-SERVER_RUNNING_BANNER = """
-+----------------------------------------+
-|             SERVER WORKS!              |
-+----------------------------------------+
-
-Visit http://{host}:{port}
-
-..........................................
-"""
