+++ src/main/wsgi.py
@@ -1,19 +1,21 @@
 def application(environ, start_response):
     status = "200 OK"
+
     headers = {
         "Content-type": "text/html",
     }
+
     payload = (
         b"<!DOCTYPE html>"
         b"<html>"
         b"<head>"
-        b"<title>Z37</title>"
+        b"<title>Alpha</title>"
         b'<meta charset="utf-8">'
         b"</head>"
         b"<body>"
-        b"<h1>Z37 study project</h1>"
+        b"<h1>Project Alpha</h1>"
         b"<hr>"
-        b"<p>This is a study project.</p>"
+        b"<p>This is a template project.</p>"
         b"</body>"
         b"</html>"
     )
