+++ src/scripts/get_db_user.py
@@ -1,4 +1,4 @@
-from utils import get_db_username
+from framework.util.db import get_db_username
 
 
 def main():
